Split cli-config.ts into folder structure (core, sites, snapshots)
