+++ apps/cli/lib/cli-config/core.ts
@@ -1,0 +1,121 @@
+import fs from 'fs';
+import path from 'path';
+import { LOCKFILE_STALE_TIME, LOCKFILE_WAIT_TIME } from '@studio/common/constants';
+import { lockFileAsync, unlockFileAsync } from '@studio/common/lib/lockfile';
+import { siteDetailsSchema } from '@studio/common/lib/site-events';
+import { snapshotSchema } from '@studio/common/types/snapshot';
+import { __ } from '@wordpress/i18n';
+import { readFile, writeFile } from 'atomically';
+import { z } from 'zod';
+import { STUDIO_CLI_HOME } from 'cli/lib/paths';
+import { LoggerError } from 'cli/logger';
+
+const siteSchema = siteDetailsSchema
+	.extend( {
+		url: z.string().optional(),
+		latestCliPid: z.number().optional(),
+	} )
+	.loose();
+
+const cliConfigWithJustVersion = z.object( {
+	version: z.number().default( 1 ),
+} );
+// IMPORTANT: Always consider that independently installed versions of the CLI (from npm) may also
+// read this file, and any updates to this schema may require updating the `version` field.
+const cliConfigSchema = cliConfigWithJustVersion.extend( {
+	sites: z.array( siteSchema ).default( () => [] ),
+	snapshots: z.array( snapshotSchema ).default( () => [] ),
+} );
+
+type CliConfig = z.infer< typeof cliConfigSchema >;
+export type SiteData = z.infer< typeof siteSchema >;
+
+const DEFAULT_CLI_CONFIG: CliConfig = {
+	version: 1,
+	sites: [],
+	snapshots: [],
+};
+
+export function getCliConfigDirectory(): string {
+	if ( process.env.E2E && process.env.E2E_CLI_CONFIG_PATH ) {
+		return process.env.E2E_CLI_CONFIG_PATH;
+	}
+
+	return STUDIO_CLI_HOME;
+}
+
+export function getCliConfigPath(): string {
+	return path.join( getCliConfigDirectory(), 'cli.json' );
+}
+
+export async function readCliConfig(): Promise< CliConfig > {
+	const configPath = getCliConfigPath();
+
+	if ( ! fs.existsSync( configPath ) ) {
+		return structuredClone( DEFAULT_CLI_CONFIG );
+	}
+
+	try {
+		const fileContent = await readFile( configPath, { encoding: 'utf8' } );
+		// eslint-disable-next-line no-var
+		var data = JSON.parse( fileContent );
+	} catch ( error ) {
+		throw new LoggerError( __( 'Failed to read CLI config file.' ), error );
+	}
+
+	try {
+		return cliConfigSchema.parse( data );
+	} catch ( error ) {
+		if ( error instanceof z.ZodError ) {
+			try {
+				cliConfigWithJustVersion.parse( data );
+			} catch ( versionError ) {
+				throw new LoggerError(
+					__(
+						'Invalid CLI config version. It looks like you have a different version of the `studio` CLI installed on your system. Please modify your $PATH environment variable to use the correct version.'
+					),
+					error
+				);
+			}
+
+			throw new LoggerError( __( 'Invalid CLI config file format.' ), error );
+		}
+
+		if ( error instanceof SyntaxError ) {
+			throw new LoggerError( __( 'CLI config file is corrupted.' ), error );
+		}
+
+		throw new LoggerError( __( 'Failed to read CLI config file.' ), error );
+	}
+}
+
+export async function saveCliConfig( config: CliConfig ): Promise< void > {
+	try {
+		config.version = 1;
+
+		const configDir = getCliConfigDirectory();
+		if ( ! fs.existsSync( configDir ) ) {
+			fs.mkdirSync( configDir, { recursive: true } );
+		}
+
+		const configPath = getCliConfigPath();
+		const fileContent = JSON.stringify( config, null, 2 ) + '\n';
+
+		await writeFile( configPath, fileContent, { encoding: 'utf8' } );
+	} catch ( error ) {
+		if ( error instanceof LoggerError ) {
+			throw error;
+		}
+		throw new LoggerError( __( 'Failed to save CLI config file' ), error );
+	}
+}
+
+const LOCKFILE_PATH = path.join( getCliConfigDirectory(), 'cli.json.lock' );
+
+export async function lockCliConfig(): Promise< void > {
+	await lockFileAsync( LOCKFILE_PATH, { wait: LOCKFILE_WAIT_TIME, stale: LOCKFILE_STALE_TIME } );
+}
+
+export async function unlockCliConfig(): Promise< void > {
+	await unlockFileAsync( LOCKFILE_PATH );
+}
